permissions frontend

+++ frontend/src/app/pages/admin/admin-meals/admin-meals.component.ts
@@ -9,6 +9,8 @@ import { MenuService } from 'src/app/services/menu.service';
 import { v4 as uuidv4 } from 'uuid';
 import  Swal  from 'sweetalert2';
 import { Alert } from 'src/assets/alert';
+import { AuthService } from '@auth0/auth0-angular';
+import { User } from 'src/app/models/user';
 
 @Component({
   selector: 'app-admin',
@@ -19,13 +21,20 @@ export class AdminMealsComponent implements OnInit {
   week : Week;
   menu : Menu;
   subscriptions : Subscription[] = [];
+  currentUser : User;
 
 
-  constructor(private sharingService : SharingService, private menuService : MenuService, private alert : Alert,private changeDetector: ChangeDetectorRef) {
+  constructor(private sharingService : SharingService, private menuService : MenuService, private alert : Alert,private changeDetector: ChangeDetectorRef, private auth: AuthService) {
 
   }
 
   ngOnInit(): void {
+
+    this.subscriptions.push(this.auth.user$.subscribe((user) => {
+      this.currentUser = new User();
+      this.currentUser.setUserFromAuthPic(user.picture);
+    }));
+
     let sub: Subscription = this.sharingService.getObservableWeek().subscribe((week : Week) => {
       this.week = week;
     });
@@ -43,21 +52,24 @@ export class AdminMealsComponent implements OnInit {
   }
 
   deleteMeal(mealId : string, day : Day) : void{
-   this.alert.showAdvancedAlert('VARNING', 'Vill du ta bort denna maträtt?', 'warning', 'Ja, ta bort', 'Avbryt').then((result) => {
-      if (result.isConfirmed) {
-        day.meals.forEach((meal, index) => {
-          if(meal._id === mealId) {
-            day.meals.splice(index, 1);
+    if(!this.currentUser.permissions.some((permission) => permission === 'admin') && !this.currentUser.menuIds.some((menuId) => menuId === this.menu._id)) {
+      this.alert.showAlert('', 'Du måste ha behörighet för att administrera denna matsedel!', 'warning');
+    } else {
+      this.alert.showAdvancedAlert('VARNING', 'Vill du ta bort denna maträtt?', 'warning', 'Ja, ta bort', 'Avbryt').then((result) => {
+          if (result.isConfirmed) {
+            day.meals.forEach((meal, index) => {
+              if(meal._id === mealId) {
+                day.meals.splice(index, 1);
+              }
+            });
+            let sub: Subscription =this.menuService.deleteMeal(this.menu._id, mealId).subscribe(() => {
+            });
+            this.subscriptions.push(sub);
+
+            this.alert.showAlert('Borttagen!', 'Vald maträtt har blivit borttagen.', 'success');
           }
-        });
-        let sub: Subscription =this.menuService.deleteMeal(this.menu._id, mealId).subscribe(() => {
-        });
-        this.subscriptions.push(sub);
-
-        this.alert.showAlert('Borttagen!', 'Vald maträtt har blivit borttagen.', 'success');
-      }
-    })
-
+        })
+    }
   }
 
   checkFoodSpec(foodSpecs: string[], wantedSpec: string): boolean {
@@ -71,57 +83,65 @@ export class AdminMealsComponent implements OnInit {
   }
 
   updateMeal(meal: Meal, day: Day, mealName : string, veg : any, hot : any, pig: any) {
-    if(mealName.length < 1) {
-      this.alert.showAlert('', 'Input för maträtten är för kort. Testa igen!', 'warning');
-    } else if (mealName.length > 85) {
-      this.alert.showAlert('', 'Input för maträtten är för långt. Testa igen!', 'warning');
+    if(!this.currentUser.permissions.some((permission) => permission === 'admin') && !this.currentUser.menuIds.some((menuId) => menuId === this.menu._id)) {
+      this.alert.showAlert('', 'Du måste ha behörighet för att administrera denna matsedel!', 'warning');
     } else {
-      meal.mealName = mealName;
-      meal.foodSpecs = [];
-        if(veg.checked) {
-          meal.foodSpecs.push(veg.value);
-        }
-        if(hot.checked) {
-          meal.foodSpecs.push(hot.value);
-        }
-        if(pig.checked) {
-          meal.foodSpecs.push(pig.value);
-        }
-        let sub: Subscription =this.menuService.updateMeal(meal, this.menu._id).subscribe(() => {
-          this.subscriptions.push(sub);
-        });
+      if(mealName.length < 1) {
+        this.alert.showAlert('', 'Input för maträtten är för kort. Testa igen!', 'warning');
+      } else if (mealName.length > 85) {
+        this.alert.showAlert('', 'Input för maträtten är för långt. Testa igen!', 'warning');
+      } else {
+        meal.mealName = mealName;
+        meal.foodSpecs = [];
+          if(veg.checked) {
+            meal.foodSpecs.push(veg.value);
+          }
+          if(hot.checked) {
+            meal.foodSpecs.push(hot.value);
+          }
+          if(pig.checked) {
+            meal.foodSpecs.push(pig.value);
+          }
+          let sub: Subscription =this.menuService.updateMeal(meal, this.menu._id).subscribe(() => {
+            this.subscriptions.push(sub);
+          });
 
-        this.alert.showAlert('Uppdaterad!', 'Maträtten har blivit uppdaterad.', 'success');
+          this.alert.showAlert('Uppdaterad!', 'Maträtten har blivit uppdaterad.', 'success');
+      }
     }
   }
 
   saveMeal(day: Day, newMealName : string, veg : any, hot : any, pig: any, form : any) {
-    if(newMealName.length < 1) {
-      this.alert.showAlert('', 'Input för maträtten är för kort. Testa igen!', 'warning');
-    } else if (newMealName.length > 85) {
-      this.alert.showAlert('', 'Input för maträtten är för långt. Testa igen!', 'warning');
+    if(!this.currentUser.permissions.some((permission) => permission === 'admin') && !this.currentUser.menuIds.some((menuId) => menuId === this.menu._id)) {
+      this.alert.showAlert('', 'Du måste ha behörighet för att administrera denna matsedel!', 'warning');
     } else {
-      let meal : Meal = new Meal();
-      meal.mealName = newMealName;
-      meal.mealDate = new Date(day.date);
-        if(veg.checked) {
-          meal.foodSpecs.push(veg.value);
-        }
-        if(hot.checked) {
-          meal.foodSpecs.push(hot.value);
-        }
-        if(pig.checked) {
-          meal.foodSpecs.push(pig.value);
-        }
-        meal._id = uuidv4();
-        day.meals.push(meal);
-        let sub : Subscription= this.menuService.postMeal(meal, this.menu._id).subscribe((mealId) => {
+      if(newMealName.length < 1) {
+        this.alert.showAlert('', 'Input för maträtten är för kort. Testa igen!', 'warning');
+      } else if (newMealName.length > 85) {
+        this.alert.showAlert('', 'Input för maträtten är för långt. Testa igen!', 'warning');
+      } else {
+        let meal : Meal = new Meal();
+        meal.mealName = newMealName;
+        meal.mealDate = new Date(day.date);
+          if(veg.checked) {
+            meal.foodSpecs.push(veg.value);
+          }
+          if(hot.checked) {
+            meal.foodSpecs.push(hot.value);
+          }
+          if(pig.checked) {
+            meal.foodSpecs.push(pig.value);
+          }
+          meal._id = uuidv4();
+          day.meals.push(meal);
+          let sub : Subscription= this.menuService.postMeal(meal, this.menu._id).subscribe((mealId) => {
 
-        });
-        this.subscriptions.push(sub);
-        form.reset();
+          });
+          this.subscriptions.push(sub);
+          form.reset();
 
-        this.alert.showAlert('Sparad!', 'Maträtten har blivit sparad.', 'success');
+          this.alert.showAlert('Sparad!', 'Maträtten har blivit sparad.', 'success');
+      }
     }
   }
 }
